+++ Services/Models/PlayCount.cs
@@ -23,14 +23,14 @@ namespace Spotcc.Services.Models
         [Indexed]
         public string Account { get; set; }
         public string ArtistName { get; set; }
-        public int Playcount { get; set; }
+        public long Playcount { get; set; }
         public int MonthlyListeners { get; set; }
         public int FollowerCount { get; set; }
 
         [Indexed]
         public DateTime Date { get; set; }
 
-        public int Delta { get; set; }
+        public long Delta { get; set; }
         public StreamType StreamType { get; set; }
     }
 }
